+++ docker/trustless-stack/config/local/kupo/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+echo "Waiting 20 seconds for cardano testnet to start..."
+sleep 20
+echo "Running kupo..."
+exec /bin/kupo "$@"