+++ docker/trustless-stack/config/local/ogmios/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+echo "Waiting 20 seconds for cardano testnet to start..."
+sleep 20
+echo "Running ogmios..."
+exec /bin/ogmios "$@"